add new datasource

+++ models/dataset.py
@@ -3,21 +3,22 @@ import urllib.request
 from .params import *
 
 class Dataset():
-    def __init__(self, params: Hyperparameters):
+    def __init__(self, params: Hyperparameters, get_data=False):
         self.params = params
         script_dir = os.path.dirname(os.path.realpath(__file__))
-        self.filename = os.path.join(script_dir, '../training_data', 'input.txt')
-        self.source_url = 'https://raw.githubusercontent.com/karpathy/char-rnn/master/data/tinyshakespeare/input.txt'
+        self.filename = os.path.join(script_dir, '../training_data', params.training_data_name)
+        self.source_url = params.training_data_url
 
-        self.data = self.get_data()
+        self.data = self.get_data(get_data)
         self.tokens = self.tokenize_data()
         self.vocab_size = len(self.tokens)
         self.encode = self.get_encoder()
         self.decode = self.get_decoder()
         self.train_data, self.test_data = self.split_data()
 
-    def get_data(self):
-        urllib.request.urlretrieve(self.source_url, self.filename)
+    def get_data(self, get_data=False):
+        if get_data or os.path.isfile(self.filename) == False:
+            urllib.request.urlretrieve(self.source_url, self.filename)
         with open(self.filename, 'r', encoding='utf-8') as f:
             data = f.read()
         return data
